requirements.txt
datasets 
fastapi 
transformers 
matplotlib
langchain 
langchain-core 
langchain-ollama 
angchain-openai 
langchain-text-splitters
langgraph 
langgraph-checkpoint 
langgraph-prebuilt 
langgraph-sdk
pypandoc 
python-docx 
pdfminer.six 
pytesseract 
pillow 
pdf2image
openai-whisper
git+https://github.com/openai/whisper.git 
ffmpeg
setuptools-rust
faster_whisper
uggingface_hub
pandoc
poppler-utils
